3 refactor project user modules validation

+++ src/address/address.service.ts
@@ -1,9 +1,16 @@
 import { Injectable, Logger } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
-import Mongose, { Model } from 'mongoose';
-import { ServiceError } from '../helpers/response/error';
+import Mongoose, { Model } from 'mongoose';
+import { toPage } from '../helpers/pagination/pagination-helper';
+import { PaginatedDto } from '../common/pagination/paginated-dto';
+import {
+  BadRequest,
+  NotFound,
+  ServerError,
+  Unauthorized,
+} from '../helpers/response/errors';
 import { ServiceResult } from '../helpers/response/result';
-import { isNearWallet } from '../utils/nearWalletValidation';
+import { isNearWallet } from '../utils/near-wallet-validation';
 import { CreateAddressDto } from './dto/create-address.dto';
 import { UpdateAddressDto } from './dto/update-address.dto';
 import { Address, AddressDocument } from './entities/address.entity';
@@ -19,10 +26,7 @@ export class AddressService {
   async create(dto: CreateAddressDto): Promise<ServiceResult<Address>> {
     try {
       if (!isNearWallet(dto.wallet)) {
-        return new ServiceResult<Address>(
-          null,
-          new ServiceError(400, `Wallet ${dto.wallet} not valid`),
-        );
+        return new BadRequest<Address>(`Wallet ${dto.wallet} not valid`);
       }
 
       const addressWalletDoc = await this.repo
@@ -30,10 +34,7 @@ export class AddressService {
         .exec();
 
       if (addressWalletDoc) {
-        return new ServiceResult<Address>(
-          null,
-          new ServiceError(400, `Wallet ${dto.wallet} isn't unique`),
-        );
+        return new BadRequest<Address>(`Wallet ${dto.wallet} isn't unique`);
       }
 
       const addressAliasDoc = await this.repo
@@ -41,48 +42,52 @@ export class AddressService {
         .exec();
 
       if (addressAliasDoc) {
-        return new ServiceResult<Address>(
-          null,
-          new ServiceError(400, `Alias ${dto.alias} isn't unique`),
-        );
+        return new BadRequest<Address>(`Alias ${dto.alias} isn't unique`);
       }
-
-      const address = new this.repo(dto);
-      const result = await address.save();
+      const result = await new this.repo(dto).save();
       return new ServiceResult<Address>(result);
     } catch (error) {
       this.logger.error('AddressService - create', error);
-      return new ServiceResult<Address>(
-        null,
-        new ServiceError(500, `Can't create address`),
-      );
+      return new ServerError<Address>(`Can't create address`);
     }
   }
 
-  async findAll(): Promise<ServiceResult<Address[]>> {
+  async findAll(
+    ownerId: Mongoose.Types.ObjectId,
+    offset?: number,
+    limit?: number,
+    alias?: string,
+  ): Promise<ServiceResult<PaginatedDto<Address>>> {
     try {
-      const addresses = await this.repo.find().populate('owner').exec();
-      return new ServiceResult<Address[]>(addresses);
+      const query = this.repo.find({ owner: ownerId });
+      const queryCount = this.repo.find({ owner: ownerId }).countDocuments();
+
+      if (alias) {
+        query.find({ alias: { $regex: alias, $options: 'i' } });
+      }
+
+      const paginatedDto = await toPage<Address>(
+        query,
+        queryCount,
+        offset,
+        limit,
+      );
+
+      return new ServiceResult<PaginatedDto<Address>>(paginatedDto);
     } catch (error) {
       this.logger.error('AddressService - findAll', error);
-      return new ServiceResult<Address[]>(
-        null,
-        new ServiceError(500, `Can't get addresses`),
-      );
+      return new ServerError<PaginatedDto<Address>>(`Can't get addresses`);
     }
   }
 
   async update(
     id: string,
-    userUid: string,
+    ownerId: string,
     updateAddressDto: UpdateAddressDto,
   ): Promise<ServiceResult<Address>> {
     try {
-      if (!Mongose.Types.ObjectId.isValid(id)) {
-        return new ServiceResult<Address>(
-          null,
-          new ServiceError(404, 'Address not found'),
-        );
+      if (!Mongoose.Types.ObjectId.isValid(id)) {
+        return new NotFound<Address>('Address not found');
       }
 
       const address = await this.repo
@@ -91,17 +96,11 @@ export class AddressService {
         .exec();
 
       if (!address) {
-        return new ServiceResult<Address>(
-          null,
-          new ServiceError(404, `Address not found`),
-        );
+        return new NotFound<Address>('Address not found');
       }
 
-      if (address.owner.uid !== userUid) {
-        return new ServiceResult<Address>(
-          null,
-          new ServiceError(401, `Unauthorized access to user address`),
-        );
+      if (address.owner._id.toString() !== ownerId) {
+        return new Unauthorized<Address>('Unauthorized access to user address');
       }
 
       const updateAddress = await this.repo.findOne({ _id: id }).exec();
@@ -111,20 +110,14 @@ export class AddressService {
       return new ServiceResult<Address>(updateAddress);
     } catch (error) {
       this.logger.error('AddressService - update', error);
-      return new ServiceResult<Address>(
-        null,
-        new ServiceError(500, `Can't update address`),
-      );
+      return new ServerError<Address>(`Can't update address`);
     }
   }
 
-  async findOne(id: string): Promise<ServiceResult<Address>> {
+  async findOne(id: string, ownerId: string): Promise<ServiceResult<Address>> {
     try {
-      if (!Mongose.Types.ObjectId.isValid(id)) {
-        return new ServiceResult<Address>(
-          null,
-          new ServiceError(404, 'Address not found'),
-        );
+      if (!Mongoose.Types.ObjectId.isValid(id)) {
+        return new NotFound<Address>('Address not found');
       }
       const address = await this.repo
         .findOne({ _id: id })
@@ -132,28 +125,24 @@ export class AddressService {
         .exec();
 
       if (!address) {
-        return new ServiceResult<Address>(
-          null,
-          new ServiceError(404, 'Address not found'),
-        );
+        return new NotFound<Address>('Address not found');
       }
+
+      if (address.owner._id.toString() !== ownerId) {
+        return new Unauthorized<Address>('Unauthorized access to user address');
+      }
+
       return new ServiceResult<Address>(address);
     } catch (error) {
       this.logger.error('AddressService - findOne', error);
-      return new ServiceResult<Address>(
-        null,
-        new ServiceError(500, `Can't get address`),
-      );
+      return new ServerError<Address>(`Can't get address`);
     }
   }
 
-  async remove(id: string): Promise<ServiceResult<Address>> {
+  async remove(id: string, ownerId: string): Promise<ServiceResult<Address>> {
     try {
-      if (!Mongose.Types.ObjectId.isValid(id)) {
-        return new ServiceResult<Address>(
-          null,
-          new ServiceError(404, `Address not found`),
-        );
+      if (!Mongoose.Types.ObjectId.isValid(id)) {
+        return new NotFound<Address>('Address not found');
       }
 
       const address = await this.repo
@@ -162,20 +151,18 @@ export class AddressService {
         .exec();
 
       if (!address) {
-        return new ServiceResult<Address>(
-          null,
-          new ServiceError(404, 'Address not found'),
-        );
+        return new NotFound<Address>('Address not found');
+      }
+
+      if (address.owner._id.toString() !== ownerId) {
+        return new Unauthorized<Address>('Unauthorized access to user address');
       }
 
       const result = await this.repo.findByIdAndDelete(id).exec();
       return new ServiceResult<Address>(result);
     } catch (error) {
       this.logger.error('AddressService - findOne', error);
-      return new ServiceResult<Address>(
-        null,
-        new ServiceError(500, `Can't remove address`),
-      );
+      return new ServerError<Address>(`Can't remove address`);
     }
   }
 }
